Incident logs (on dashboard only for now)
Util.log with type > 1 (info) will now create a new incident log.
All incidents will be available on a dedicated system page, that
will be added in the next commits. But for now, they are only available
on the Dashboard page.

Also added rights for viewing EDE status on dashboard.